requirements.txt
telethon
google-auth-oauthlib
google-auth-httplib2
google-auth
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
httplib2
requests
asyncio
ffmpeg
